Updated the Login React Component ES5 code to ES6
Added necessary imports and changed all the old custom functions to
ES6 class member functions.
Added a constructor to replace the getInitialState function and added
manual bindings for the 'this' variable in the custom onStart and
onStop functions that rely on updating the activeDrags variable (ES6
classes disallows autobinding).
Modified getDefaultProps to a static defaultProps function.
Modified the component attachment to an export default.

Still need to fix the event handlers that rely on functions that live
in external scripts (clean() and connect() are undefined).