Added logic for visual drag and drop cues
Added function constants drop and canDrop in the collecting function
fileTarget to return props necessary to render style changes depending
on dropping.

Added the endDrag function to the fileSource collecting function to
print which file was dropped onto which file.

Added injectDropProps for the DropTarget.

Updated the props destructuring and implemented logic to render the
file background color and font size depending on drag/drop state.